stores: address PJ's comments
